- constify - Move findBaseState into seperate DFA method - Prepare -wb support # Basically we can support -wb if we find a good way to reduce the number # of Go's to those that have their high byte set.

+++ code.h
@@ -17,7 +17,7 @@ public:
 	const State     *on;
 	const BitMap    *next;
 	uint            i;
-	uchar           m;
+	uint            m;
 
 public:
 	static const BitMap *find(const Go*, const State*);
